[move]: move User package into domain package
