+++ restApiServer/src/main/java/com/esctb/restapiserver/domain/user/controller/UserApi.java
@@ -1,0 +1,4 @@
+package com.esctb.restapiserver.domain.user.controller;
+
+public class UserApi {
+}
