+++ restApiServer/src/main/java/com/esctb/restapiserver/domain/user/entity/Address.java
@@ -1,4 +1,4 @@
-package com.esctb.restapiserver.user.entity;
+package com.esctb.restapiserver.domain.user.entity;
 
 import lombok.Getter;
