+++ restApiServer/src/main/java/com/esctb/restapiserver/domain/user/entity/UserImage.java
@@ -1,4 +1,4 @@
-package com.esctb.restapiserver.user.entity;
+package com.esctb.restapiserver.domain.user.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,11 +10,15 @@ import javax.persistence.*;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Area {
+@Table(name="user_image")
+public class UserImage {
+
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="area_id")
+    @Column(name="user_image_id")
     private Long id;
 
-    @Column(name = "area_name")
-    private String areaName;
+    @Column(name = "image_path")
+    private String imagePath;
+
+    //private String userId;
 }
