+++ restApiServer/src/main/java/com/esctb/restapiserver/domain/user/entity/enums/SuedYn.java
@@ -1,0 +1,5 @@
+package com.esctb.restapiserver.domain.user.entity.enums;
+
+public enum SuedYn {
+    Y, N
+}
