+++ restApiServer/src/main/java/com/esctb/restapiserver/domain/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.esctb.restapiserver.domain.user.service;
+
+public class UserService {
+}
