--- restApiServer/src/main/java/com/esctb/restapiserver/user/controller/UserApi.java
@@ -1,4 +1,0 @@
-package com.esctb.restapiserver.user.controller;
-
-public class UserApi {
-}
