--- restApiServer/src/main/java/com/esctb/restapiserver/user/entity/enums/SuedYn.java
@@ -1,5 +1,0 @@
-package com.esctb.restapiserver.user.entity.enums;
-
-public enum SuedYn {
-    Y, N
-}
