--- restApiServer/src/main/java/com/esctb/restapiserver/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.esctb.restapiserver.user.service;
-
-public class UserService {
-}
